requirements.txt
amqp==5.1.1
asgiref==3.8.1
async-timeout==5.0.0
attrs==25.1.0
billiard==4.2.0
black==25.1.0
boto3==1.36.11
botocore==1.36.11
celery==5.4.0
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.10
cron_descriptor==1.4.0
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.3.0
dj-stripe==2.9.0
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-filter==24.3
django-filter-stubs==0.1.3
django-phonenumber-field==8.0.0
django-storages==1.14.2
django-stubs==5.1.2
django-stubs-ext==5.1.2
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-stubs==3.14.3
djoser==2.3.1
drf-yasg==1.21.8
drf-yasg-stubs==0.1.4
exceptiongroup==1.1.0
factory-boy==3.3.0
Faker==35.2.0
flake8==7.1.1
freezegun==1.5.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.0
pillow==11.1.0
pipdeptree==2.13.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg==3.1.12
pycodestyle==2.12.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.0.9
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-freezegun==0.4.2
pytest-mock==3.12.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.7
s3transfer==0.11.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.5.3
stripe==11.5.0
tomli==2.0.1
types-pytz==2025.1.0.20250204
types-PyYAML==6.0.12.12
types-requests==2.32.0.20241016
types-stripe==3.5.2.20240106
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.5
wrapt==1.14.1
